refactor(cli): comprehensive optimization and structural overhaul
This commit consolidates the work to modularize and optimize the Kates CLI Go codebase:
- Resolved various bugs including error handling, os.Exit usages, deprecated APIs, and dynamic cluster domain detection.
- Extracted internal/kubectl package to centralize and replace raw exec.Command calls.
- Added internal/helm package to replace raw exec.Command helm calls.
- Created JSON output middleware to eliminate repetitive output guards.
- Consolidated local lipgloss styles to use centralized output styles and theme tokens.
- Split the monolithic deploy orchestrator into modular files and extracted wait logic.
- Split the massive security command into domain-specific files.
- Embedded Go data structs into YAML, loaded via //go:embed to reduce source code size.
- Extracted shared helpers including column width formatting, cluster domain resolution, component structs, and group names.
