+++ cli/cmd/connect.go
@@ -9,8 +9,9 @@ import (
 	"sort"
 	"strconv"
 	"strings"
-	"text/tabwriter"
 
+	"github.com/klster/kates-cli/internal/kubectl"
+	"github.com/klster/kates-cli/output"
 	"github.com/spf13/cobra"
 )
 
@@ -26,15 +27,15 @@ var connectStatusCmd = &cobra.Command{
 	Use:   "status",
 	Short: "Show Connect cluster status",
 	RunE: func(cmd *cobra.Command, args []string) error {
-		argsCmd := []string{"get", "kafkaconnect", "-n", connectNamespace}
-		if outputMode == "json" {
-			argsCmd = append(argsCmd, "-o", "json")
-		}
-		out, err := exec.CommandContext(context.Background(), "kubectl", argsCmd...).Output()
+		kc := kubectl.New(connectNamespace)
+		ctx := context.Background()
+		out, err := kc.Output(ctx, "get", "kafkaconnect", "-n", connectNamespace)
 		if err != nil {
 			return fmt.Errorf("failed to get kafkaconnect: %w", err)
 		}
-		fmt.Print(string(out))
+		output.Render(outputMode == "json", string(out), func() {
+			fmt.Print(string(out))
+		})
 		return nil
 	},
 }
@@ -43,16 +44,10 @@ var connectConnectorsCmd = &cobra.Command{
 	Use:   "connectors",
 	Short: "List all KafkaConnector CRs",
 	RunE: func(cmd *cobra.Command, args []string) error {
-		if outputMode == "json" {
-			out, err := exec.CommandContext(context.Background(), "kubectl", "get", "kafkaconnector", "-n", connectNamespace, "-o", "json").Output()
-			if err != nil {
-				return fmt.Errorf("failed to get kafkaconnector: %w", err)
-			}
-			fmt.Print(string(out))
-			return nil
-		}
+		kc := kubectl.New(connectNamespace)
+		ctx := context.Background()
 
-		out, err := exec.CommandContext(context.Background(), "kubectl", "get", "kafkaconnector", "-n", connectNamespace, "-o", "json").Output()
+		out, err := kc.Output(ctx, "get", "kafkaconnector", "-n", connectNamespace, "-o", "json")
 		if err != nil {
 			return fmt.Errorf("failed to get kafkaconnector: %w", err)
 		}
@@ -60,7 +55,7 @@ var connectConnectorsCmd = &cobra.Command{
 		var list kafkaConnectorList
 		if err := json.Unmarshal(out, &list); err != nil {
 			// Keep old behavior as fallback when JSON parsing fails.
-			fallback, fbErr := exec.CommandContext(context.Background(), "kubectl", "get", "kafkaconnector", "-n", connectNamespace).Output()
+			fallback, fbErr := kc.Output(ctx, "get", "kafkaconnector", "-n", connectNamespace)
 			if fbErr != nil {
 				return fmt.Errorf("failed to parse connector output (%v) and fallback list failed: %w", err, fbErr)
 			}
@@ -68,46 +63,48 @@ var connectConnectorsCmd = &cobra.Command{
 			return nil
 		}
 
-		w := tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', 0)
-		fmt.Fprintln(w, "NAME\tCLUSTER\tCONNECTOR CLASS\tMAX TASKS\tREADY\tRUNNING TASKS\tTASK STATES")
-		for _, item := range list.Items {
-			running := 0
-			statesSet := map[string]struct{}{}
-			for _, task := range item.Status.ConnectorStatus.Tasks {
-				if strings.EqualFold(task.State, "RUNNING") {
-					running++
+		output.Render(outputMode == "json", list, func() {
+			rows := make([][]string, 0, len(list.Items))
+			for _, item := range list.Items {
+				running := 0
+				statesSet := map[string]struct{}{}
+				for _, task := range item.Status.ConnectorStatus.Tasks {
+					if strings.EqualFold(task.State, "RUNNING") {
+						running++
+					}
+					if task.State != "" {
+						statesSet[task.State] = struct{}{}
+					}
 				}
-				if task.State != "" {
-					statesSet[task.State] = struct{}{}
+				states := make([]string, 0, len(statesSet))
+				for s := range statesSet {
+					states = append(states, s)
 				}
-			}
-			states := make([]string, 0, len(statesSet))
-			for s := range statesSet {
-				states = append(states, s)
-			}
-			sort.Strings(states)
-			taskStates := "-"
-			if len(states) > 0 {
-				taskStates = strings.Join(states, ",")
-			}
+				sort.Strings(states)
+				taskStates := "-"
+				if len(states) > 0 {
+					taskStates = strings.Join(states, ",")
+				}
 
-			maxTasks := item.Spec.TasksMax
-			runningTasks := strconv.Itoa(running)
-			if maxTasks > 0 {
-				runningTasks = fmt.Sprintf("%d/%d", running, maxTasks)
-			}
+				maxTasks := item.Spec.TasksMax
+				runningTasks := strconv.Itoa(running)
+				if maxTasks > 0 {
+					runningTasks = fmt.Sprintf("%d/%d", running, maxTasks)
+				}
 
-			fmt.Fprintf(w, "%s\t%s\t%s\t%d\t%s\t%s\t%s\n",
-				item.Metadata.Name,
-				item.Metadata.Labels["strimzi.io/cluster"],
-				item.Spec.Class,
-				maxTasks,
-				isReadyConditionTrue(item.Status.Conditions),
-				runningTasks,
-				taskStates,
-			)
-		}
-		return w.Flush()
+				rows = append(rows, []string{
+					item.Metadata.Name,
+					item.Metadata.Labels["strimzi.io/cluster"],
+					item.Spec.Class,
+					fmt.Sprintf("%d", maxTasks),
+					isReadyConditionTrue(item.Status.Conditions),
+					runningTasks,
+					taskStates,
+				})
+			}
+			output.Table([]string{"Name", "Cluster", "Connector Class", "Max Tasks", "Ready", "Running Tasks", "Task States"}, rows)
+		})
+		return nil
 	},
 }
 
@@ -116,17 +113,21 @@ var connectConnectorCmd = &cobra.Command{
 	Short: "Describe a connector",
 	Args:  cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
+		kc := kubectl.New(connectNamespace)
+		ctx := context.Background()
 		argsCmd := []string{"get", "kafkaconnector", args[0], "-n", connectNamespace}
-		if outputMode == "json" {
-			argsCmd = append(argsCmd, "-o", "json")
-		} else {
+		if outputMode != "json" {
 			argsCmd = append(argsCmd, "-o", "yaml")
+		} else {
+			argsCmd = append(argsCmd, "-o", "json")
 		}
-		out, err := exec.CommandContext(context.Background(), "kubectl", argsCmd...).Output()
+		out, err := kc.Output(ctx, argsCmd...)
 		if err != nil {
 			return fmt.Errorf("failed to describe connector %s: %w", args[0], err)
 		}
-		fmt.Print(string(out))
+		output.Render(outputMode == "json", string(out), func() {
+			fmt.Print(string(out))
+		})
 		return nil
 	},
 }
@@ -136,45 +137,33 @@ var connectTasksCmd = &cobra.Command{
 	Short: "Show task-level status for a connector",
 	Args:  cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
-		out, err := exec.CommandContext(context.Background(), "kubectl", "get", "kafkaconnector", args[0], "-n", connectNamespace, "-o", "json").Output()
-		if err != nil {
+		kc := kubectl.New(connectNamespace)
+		ctx := context.Background()
+
+		var connector kafkaConnector
+		if err := kc.JSON(ctx, &connector, "get", "kafkaconnector", args[0], "-n", connectNamespace); err != nil {
 			return fmt.Errorf("failed to get tasks for connector %s: %w", args[0], err)
 		}
 
-		var connector kafkaConnector
-		if err := json.Unmarshal(out, &connector); err != nil {
-			return fmt.Errorf("failed to parse connector %s status: %w", args[0], err)
-		}
-
 		tasks := connector.Status.ConnectorStatus.Tasks
-		if outputMode == "json" {
-			b, err := json.Marshal(tasks)
-			if err != nil {
-				return fmt.Errorf("failed to marshal tasks: %w", err)
+		output.Render(outputMode == "json", tasks, func() {
+			if len(tasks) == 0 {
+				output.Hint(fmt.Sprintf("No tasks found for connector %s.", args[0]))
+				return
 			}
-			fmt.Println(string(b))
-			return nil
-		}
 
-		if len(tasks) == 0 {
-			connectorState := connector.Status.ConnectorStatus.Connector.State
-			if connectorState == "" {
-				connectorState = "UNKNOWN"
+			output.Banner(fmt.Sprintf("Connector: %s", args[0]), fmt.Sprintf("%d tasks", len(tasks)))
+			rows := make([][]string, 0, len(tasks))
+			for _, task := range tasks {
+				rows = append(rows, []string{
+					fmt.Sprintf("%d", task.ID),
+					task.State,
+					task.WorkerID,
+					task.Version,
+				})
 			}
-			fmt.Printf("No tasks currently reported for connector %s (connector state: %s, ready: %s).\n",
-				args[0],
-				connectorState,
-				isReadyConditionTrue(connector.Status.Conditions),
-			)
-			return nil
-		}
-
-		w := tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', 0)
-		fmt.Fprintln(w, "ID\tSTATE\tWORKER ID\tVERSION")
-		for _, task := range tasks {
-			fmt.Fprintf(w, "%d\t%s\t%s\t%s\n", task.ID, task.State, task.WorkerID, task.Version)
-		}
-		w.Flush()
+			output.Table([]string{"Task ID", "State", "Worker ID", "Version"}, rows)
+		})
 		return nil
 	},
 }
@@ -184,9 +173,11 @@ var connectRestartCmd = &cobra.Command{
 	Short: "Restart a connector",
 	Args:  cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
-		out, err := exec.CommandContext(context.Background(), "kubectl", "annotate", "kafkaconnector", args[0], "-n", connectNamespace, "strimzi.io/restart=true", "--overwrite").CombinedOutput()
+		kc := kubectl.New(connectNamespace)
+		ctx := context.Background()
+		_, err := kc.Run(ctx, "annotate", "kafkaconnector", args[0], "-n", connectNamespace, "strimzi.io/restart=true", "--overwrite")
 		if err != nil {
-			return fmt.Errorf("failed to restart connector %s: %s", args[0], string(out))
+			return fmt.Errorf("failed to restart connector %s: %w", args[0], err)
 		}
 		fmt.Printf("Connector %s restart triggered.\n", args[0])
 		return nil
@@ -198,9 +189,11 @@ var connectPauseCmd = &cobra.Command{
 	Short: "Pause a connector",
 	Args:  cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
-		out, err := exec.CommandContext(context.Background(), "kubectl", "patch", "kafkaconnector", args[0], "-n", connectNamespace, "--type=merge", "-p", `{"spec":{"state":"paused"}}`).CombinedOutput()
+		kc := kubectl.New(connectNamespace)
+		ctx := context.Background()
+		_, err := kc.Run(ctx, "patch", "kafkaconnector", args[0], "-n", connectNamespace, "--type=merge", "-p", `{"spec":{"state":"paused"}}`)
 		if err != nil {
-			return fmt.Errorf("failed to pause connector %s: %s", args[0], string(out))
+			return fmt.Errorf("failed to pause connector %s: %w", args[0], err)
 		}
 		fmt.Printf("Connector %s paused.\n", args[0])
 		return nil
@@ -212,9 +205,11 @@ var connectResumeCmd = &cobra.Command{
 	Short: "Resume a connector",
 	Args:  cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
-		out, err := exec.CommandContext(context.Background(), "kubectl", "patch", "kafkaconnector", args[0], "-n", connectNamespace, "--type=merge", "-p", `{"spec":{"state":"running"}}`).CombinedOutput()
+		kc := kubectl.New(connectNamespace)
+		ctx := context.Background()
+		_, err := kc.Run(ctx, "patch", "kafkaconnector", args[0], "-n", connectNamespace, "--type=merge", "-p", `{"spec":{"state":"running"}}`)
 		if err != nil {
-			return fmt.Errorf("failed to resume connector %s: %s", args[0], string(out))
+			return fmt.Errorf("failed to resume connector %s: %w", args[0], err)
 		}
 		fmt.Printf("Connector %s resumed.\n", args[0])
 		return nil
@@ -225,7 +220,9 @@ var connectPluginsCmd = &cobra.Command{
 	Use:   "plugins",
 	Short: "List installed connector plugins",
 	RunE: func(cmd *cobra.Command, args []string) error {
-		out, err := exec.CommandContext(context.Background(), "kubectl", "get", "kafkaconnect", "-n", connectNamespace, "-o", "jsonpath={.items[0].status.connectorPlugins}").Output()
+		kc := kubectl.New(connectNamespace)
+		ctx := context.Background()
+		out, err := kc.Output(ctx, "get", "kafkaconnect", "-n", connectNamespace, "-o", "jsonpath={.items[0].status.connectorPlugins}")
 		if err != nil {
 			return fmt.Errorf("failed to get plugins (ensure a KafkaConnect cluster is running): %w", err)
 		}
@@ -265,10 +262,12 @@ var connectRestartTaskCmd = &cobra.Command{
 	Short: "Restart a specific connector task",
 	Args:  cobra.ExactArgs(2),
 	RunE: func(cmd *cobra.Command, args []string) error {
+		kc := kubectl.New(connectNamespace)
+		ctx := context.Background()
 		annotation := fmt.Sprintf("strimzi.io/restart-task=%s", args[1])
-		out, err := exec.CommandContext(context.Background(), "kubectl", "annotate", "kafkaconnector", args[0], "-n", connectNamespace, annotation, "--overwrite").CombinedOutput()
+		_, err := kc.Run(ctx, "annotate", "kafkaconnector", args[0], "-n", connectNamespace, annotation, "--overwrite")
 		if err != nil {
-			return fmt.Errorf("failed to restart task %s on connector %s: %s", args[1], args[0], string(out))
+			return fmt.Errorf("failed to restart task %s on connector %s: %w", args[1], args[0], err)
 		}
 		fmt.Printf("Task %s on connector %s restart triggered.\n", args[1], args[0])
 		return nil
@@ -280,6 +279,8 @@ var connectConfigCmd = &cobra.Command{
 	Short: "Show connector configuration",
 	Args:  cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
+		kc := kubectl.New(connectNamespace)
+		ctx := context.Background()
 		format := "jsonpath={.spec.config}"
 		if outputMode == "json" {
 			format = "-o=json"
@@ -292,7 +293,7 @@ var connectConfigCmd = &cobra.Command{
 		} else {
 			cmdArgs = append(cmdArgs, format)
 		}
-		out, err := exec.CommandContext(context.Background(), "kubectl", cmdArgs...).Output()
+		out, err := kc.Output(ctx, cmdArgs...)
 		if err != nil {
 			return fmt.Errorf("failed to get config for connector %s: %w", args[0], err)
 		}
@@ -306,9 +307,11 @@ var connectDeleteCmd = &cobra.Command{
 	Short: "Delete a connector",
 	Args:  cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
-		out, err := exec.CommandContext(context.Background(), "kubectl", "delete", "kafkaconnector", args[0], "-n", connectNamespace).CombinedOutput()
+		kc := kubectl.New(connectNamespace)
+		ctx := context.Background()
+		_, err := kc.Run(ctx, "delete", "kafkaconnector", args[0], "-n", connectNamespace)
 		if err != nil {
-			return fmt.Errorf("failed to delete connector %s: %s", args[0], string(out))
+			return fmt.Errorf("failed to delete connector %s: %w", args[0], err)
 		}
 		fmt.Printf("Connector %s deleted.\n", args[0])
 		return nil
@@ -320,9 +323,11 @@ var connectScaleCmd = &cobra.Command{
 	Short: "Scale Kafka Connect workers",
 	Args:  cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
+		kc := kubectl.New(connectNamespace)
+		ctx := context.Background()
 		patch := fmt.Sprintf(`{"spec":{"replicas":%s}}`, args[0])
 		// Find the KafkaConnect resource name
-		nameOut, err := exec.CommandContext(context.Background(), "kubectl", "get", "kafkaconnect", "-n", connectNamespace, "-o", "jsonpath={.items[0].metadata.name}").Output()
+		nameOut, err := kc.Output(ctx, "get", "kafkaconnect", "-n", connectNamespace, "-o", "jsonpath={.items[0].metadata.name}")
 		if err != nil {
 			return fmt.Errorf("failed to find KafkaConnect resource: %w", err)
 		}
@@ -330,9 +335,9 @@ var connectScaleCmd = &cobra.Command{
 		if name == "" {
 			return fmt.Errorf("no KafkaConnect resource found in namespace %s", connectNamespace)
 		}
-		out, patchErr := exec.CommandContext(context.Background(), "kubectl", "patch", "kafkaconnect", name, "-n", connectNamespace, "--type=merge", "-p", patch).CombinedOutput()
+		_, patchErr := kc.Run(ctx, "patch", "kafkaconnect", name, "-n", connectNamespace, "--type=merge", "-p", patch)
 		if patchErr != nil {
-			return fmt.Errorf("failed to scale connect: %s", string(out))
+			return fmt.Errorf("failed to scale connect: %w", patchErr)
 		}
 		fmt.Printf("Kafka Connect %s scaled to %s replicas.\n", name, args[0])
 		return nil
@@ -367,8 +372,9 @@ func detectConnectNamespace() string {
 	}
 
 	// Auto-detect from cluster: find the namespace of any KafkaConnect CR
-	out, err := exec.Command("kubectl", "get", "kafkaconnect", "-A",
-		"-o", "jsonpath={.items[0].metadata.namespace}").Output()
+	kc := kubectl.New("")
+	out, err := kc.Output(context.Background(), "get", "kafkaconnect", "-A",
+		"-o", "jsonpath={.items[0].metadata.namespace}")
 	if err == nil {
 		ns := strings.TrimSpace(string(out))
 		if ns != "" {
