+++ cli/cmd/helm_test_cmd.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"bytes"
+	"context"
 	"encoding/json"
 	"fmt"
 	"os"
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"github.com/go-pdf/fpdf"
+	"github.com/klster/kates-cli/internal/helm"
 	"github.com/klster/kates-cli/output"
 	"github.com/spf13/cobra"
 )
@@ -854,13 +855,12 @@ func exportHelmTestPDF(suite HelmTestSuiteResult, filename string) error {
 // ---------------------------------------------------------------------------
 
 func runHelmCmd(args ...string) (string, error) {
-	cmd := exec.Command("helm", args...)
-	var stdout, stderr bytes.Buffer
-	cmd.Stdout = &stdout
-	cmd.Stderr = &stderr
-	err := cmd.Run()
-	combined := stdout.String() + stderr.String()
-	return strings.TrimSpace(combined), err
+	hc := helm.New("")
+	out, err := hc.Run(context.Background(), args...)
+	if err != nil {
+		return out + "\n" + err.Error(), err
+	}
+	return out, nil
 }
 
 func currentKubeContext() string {
